refactor change isAutherized name and change project architecture
