+++ src/components/header/header/Header.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Box, Toolbar, AppBar } from '@mui/material';
-import { HeaderNav } from '../index';
+import { HeaderNav } from '../headerNav';
+import { HeaderMenu } from '../headerMenu';
 import './Header.sass';
-import HeaderMenu from '../headerMenu/HeaderMenu';
 
 class Header extends React.PureComponent {
   render() {
