+++ src/components/header/header/Header.sass
@@ -1,4 +1,4 @@
-@import '../../colors'
+@import '../../../colors'
 
 .header
   display: flex
