+++ src/components/header/headerMenu/HeaderMenu.jsx
@@ -1,21 +1,21 @@
-import React from "react";
+import React from 'react';
 import { Button, ButtonGroup } from '@mui/material';
-import { getFieldLabel } from '../../utils';
+import { getFieldLabel } from '../../../utils';
 
-class HeaderMenu extends React.Component{
-  constructor(props){
+class HeaderMenu extends React.Component {
+  constructor(props) {
     super(props);
-    this.state={};
+    this.state = {};
   }
 
-  render(){
-    return(
-      <ButtonGroup variant="text" aria-label="text button group" >
+  render() {
+    return (
+      <ButtonGroup variant="text" aria-label="text button group">
         <Button>{getFieldLabel('header.menu.button.internships')}</Button>
         <Button>{getFieldLabel('header.menu.button.candidates')}</Button>
         <Button>{getFieldLabel('header.menu.button.letters')}</Button>
       </ButtonGroup>
-    )
+    );
   }
 }
