+++ src/components/header/headerNav/HeaderNav.sass
@@ -1,4 +1,4 @@
-@import '../../colors'
+@import '../../../colors'
 
 .headerNav
   width: 10%
