+++ src/components/header/index.js
@@ -1,1 +1,3 @@
-export { default as Header } from './Header';
+export { HeaderMenu } from './headerMenu';
+export { Header } from './header';
+export { HeaderNav } from './headerNav';
