+++ src/components/index.js
@@ -1,8 +1,4 @@
 export * from './footer';
-export * from './headerNav';
 export * from './header';
-export * from './sidebar';
-export * from './sidebarHeader';
 export * from './main';
-export * from './mainContentWrapper';
 export * from './utils';
