+++ src/components/main/main/index.js
@@ -1,0 +1,1 @@
+export { default as Main } from './Main';
