+++ src/components/main/mainContentWrapper/MainContentWrapper.sass
@@ -1,16 +1,13 @@
-@import '../../colors'
-    
-.main
-  padding: 0
-  background-color: $common-background 
-  display: flex
-  overflow: hidden
-  flex-direction: column
-  flex-grow: 1 
-  height: 100vh
+@import '../../../colors'
+
+.pageGap
+  margin-top: 4px
+  margin-bottom: auto
+
 
 #content
   margin-top: 70px
+  height: 100%
   display: flex
   flex-direction: column
   overflow-y: auto
@@ -22,3 +19,4 @@
   overflow-y: auto
   background: $common-elements
   box-shadow: 0px 5.2px 10.4px rgba(0, 0, 0, 0.3)
+
