--- src/components/sidebar/Sidebar.sass
@@ -1,11 +1,0 @@
-@import '../../colors'
-
-#drawer
-  width: 300px
-  & .MuiDrawer-paper
-      width: 300px
-      box-sizing: border-box
-      background-color: $primary
-      color: $common-elements
-      
-
