--- src/components/sidebar/index.js
@@ -1,1 +1,0 @@
-export { default as Sidebar } from './Sidebar';
