--- src/components/sidebarHeader/SidebarHeader.sass
@@ -1,11 +1,0 @@
-@import '../../colors'
-.sidebarHeader
-  display: flex
-  align-items: center
-  padding: 0 4px 0 4px
-  background-color: $sidebar-header
-  color: $sidebar-header-title
-  flex-direction: row
-  justify-content: space-between
-  height: 64px
-
