--- src/components/sidebarHeader/index.js
@@ -1,1 +1,0 @@
-export { default as SidebarHeader } from './SidebarHeader';
